[Bindless][NFC] Move common utility functions into vulkan_common.hpp
* move utilities that can be used by sampled and unsampled images tests
  under namespace (vk)util to header vulkan_common.hpp.
